installer: script around the futile GPT/UEFI/ZFS situation
bsdinstall does not support hybrid boot which works fine
for us since 2016.  This situation of no hybrid boot might
also break ZFS boot on UEFI as indicated by the current code
in is_fs_bootable().

Changing bsdinstall wizard is also futile as it believes that
a boot partition is to be handled special and there is only
one of those which is not the case in hybrid boot.